third_party/blink/web_tests/external/wpt/css/css-text-decor/parsing/text-decoration-trim-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-decoration-trim'] = "0px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "0" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "0px 0px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "0.5em" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "-1em" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "-1ch -1ch" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "1px 2px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "1ch -1ch" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "calc(1em / 4) calc(-1ch)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-decoration-trim'] = "auto" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
